feat: publicacion de solicitud

+++ lib/features/requests/domain/useCase/deleted_request_use_case.dart
@@ -1,17 +1,17 @@
+import 'package:fpdart/fpdart.dart';
+import 'package:servi_pro/core/errors/failures.dart';
 import 'package:servi_pro/features/requests/domain/repository/request_repository.dart';
 
 class DeletedRequestUseCase {
   final RequestRepository repository;
 
   DeletedRequestUseCase({required this.repository});
-  Future<void> call(String requestId) async {
-    
-    try {
-      await repository.deleteRequest(requestId);
-      
-    } catch (e) {
-      throw Exception("error al cancelar solicitud: $e");
+
+  Future<Either<Failure, Unit>> call(String requestId) async {
+    if (requestId.trim().isEmpty) {
+      return left(const ValidationFailure(message: 'ID de solicitud inválido'));
     }
-    
+
+    return await repository.deleteRequest(requestId);
   }
 }
